# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Zambia Multi-Tenant Enterprise Credit Scoring Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /consent/grant:
    post:
      operationId: grantConsent
      tags: [consent]
      summary: Grant data sharing consent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentGrantRequest"
      responses:
        "200":
          description: Consent granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentResponse"
  /consent/revoke:
    post:
      operationId: revokeConsent
      tags: [consent]
      summary: Revoke data sharing consent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentRevokeRequest"
      responses:
        "200":
          description: Consent revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentResponse"
  /consent/status:
    get:
      operationId: getConsentStatus
      tags: [consent]
      summary: Get consent status for current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Consent status list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentStatusResponse"
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants (super admin only)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantListResponse"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant details
      security:
        - BearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    put:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant
      security:
        - BearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantRequest"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete tenant
      security:
        - BearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tenants/{tenantId}/api-key:
    post:
      operationId: regenerateTenantApiKey
      tags: [tenants]
      summary: Regenerate API key for tenant
      security:
        - BearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: New API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyResponse"
...
  /analytics/tenant:
    get:
      operationId: getTenantAnalytics
      tags: [analytics]
      summary: Get tenant-specific analytics
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Tenant analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantAnalytics"
  /integrations/bank/fetch:
    post:
      operationId: fetchBankData
      tags: [integrations]
      summary: Fetch bank data for a customer (mock)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchDataRequest"
      responses:
        "200":
          description: Bank data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankDataResponse"
  /integrations/mno/fetch:
    post:
      operationId: fetchMnoData
      tags: [integrations]
      summary: Fetch mobile money data from MNO (mock)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchDataRequest"
      responses:
        "200":
          description: MNO mobile money data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MnoDataResponse"
  /integrations/mfi/fetch:
    post:
      operationId: fetchMfiData
      tags: [integrations]
      summary: Fetch MFI loan data (mock)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchDataRequest"
      responses:
        "200":
          description: MFI loan data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MfiDataResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        tenantCode:
          type: string
          description: Optional tenant code for tenant users
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        expiresAt:
          type: string
          format: date-time
      required:
        - token
        - user
        - expiresAt
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, tenant_admin, tenant_user, customer]
        tenantId:
          type: string
          nullable: true
        tenantName:
          type: string
          nullable: true
      required:
        - id
        - email
        - name
        - role
...
    LoanRecord:
      type: object
      properties:
        id:
          type: string
        institution:
          type: string
        institutionType:
          type: string
          enum: [bank, mfi, mno]
        amount:
          type: number
        currency:
          type: string
        outstandingBalance:
          type: number
        status:
          type: string
          enum: [active, defaulted, closed, written_off]
        disbursedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
          nullable: true
        interestRate:
          type: number
        missedPayments:
          type: integer
      required:
        - id
        - institution
        - institutionType
        - amount
        - currency
        - outstandingBalance
        - status
        - disbursedAt
        - interestRate
        - missedPayments
...
    ConsentGrantRequest:
      type: object
      properties:
        dataTypes:
          type: array
          items:
            type: string
            enum: [bank_data, mobile_money, mfi_loans, credit_history, personal_info]
        tenantId:
          type: string
          description: Tenant requesting consent
        expiresAt:
          type: string
          format: date-time
          nullable: true
      required:
        - dataTypes
    ConsentRevokeRequest:
      type: object
      properties:
        dataTypes:
          type: array
          items:
            type: string
        tenantId:
          type: string
          nullable: true
      required:
        - dataTypes
    ConsentResponse:
      type: object
      properties:
        success:
          type: boolean
        consents:
          type: array
          items:
            $ref: "#/components/schemas/ConsentRecord"
      required:
        - success
        - consents
    ConsentRecord:
      type: object
      properties:
        id:
          type: string
        dataType:
          type: string
        tenantId:
          type: string
          nullable: true
        tenantName:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, revoked]
        grantedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - dataType
        - status
        - grantedAt
    ConsentStatusResponse:
      type: object
      properties:
        customerId:
          type: string
        consents:
          type: array
          items:
            $ref: "#/components/schemas/ConsentRecord"
      required:
        - customerId
        - consents
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [bank, mfi, fintech, mno]
        status:
          type: string
          enum: [active, suspended, inactive]
        contactEmail:
          type: string
        apiCallsThisMonth:
          type: integer
        totalQueries:
          type: integer
        createdAt:
          type: string
          format: date-time
        settings:
          $ref: "#/components/schemas/TenantSettings"
      required:
        - id
        - name
        - code
        - type
        - status
        - contactEmail
        - apiCallsThisMonth
        - totalQueries
        - createdAt
    TenantSettings:
      type: object
      properties:
        scoringModel:
          type: string
          enum: [standard, conservative, aggressive]
        maxLoanAmount:
          type: number
        requireConsent:
          type: boolean
        allowedDataTypes:
          type: array
          items:
            type: string
      required:
        - scoringModel
        - maxLoanAmount
        - requireConsent
        - allowedDataTypes
    TenantListResponse:
      type: object
      properties:
        tenants:
          type: array
          items:
            $ref: "#/components/schemas/Tenant"
        total:
          type: integer
      required:
        - tenants
        - total
    CreateTenantRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [bank, mfi, fintech, mno]
        contactEmail:
          type: string
        adminName:
          type: string
        adminPassword:
          type: string
        settings:
          $ref: "#/components/schemas/TenantSettings"
      required:
        - name
        - code
        - type
        - contactEmail
        - adminName
        - adminPassword
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, suspended, inactive]
        contactEmail:
          type: string
        settings:
          $ref: "#/components/schemas/TenantSettings"
    ApiKeyResponse:
      type: object
      properties:
        apiKey:
          type: string
        tenantId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - apiKey
        - tenantId
        - createdAt
...
    TenantAnalytics:
      type: object
      properties:
        tenantId:
          type: string
        tenantName:
          type: string
        queriesThisMonth:
          type: integer
        totalQueries:
          type: integer
        avgCreditScore:
          type: number
        approvalRate:
          type: number
        riskDistribution:
          $ref: "#/components/schemas/RiskDistribution"
        recentQueries:
          type: array
          items:
            $ref: "#/components/schemas/RecentQuery"
        queryTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
      required:
        - tenantId
        - tenantName
        - queriesThisMonth
        - totalQueries
        - avgCreditScore
        - approvalRate
        - riskDistribution
        - recentQueries
        - queryTrend
    RiskDistribution:
      type: object
      properties:
        low:
          type: integer
        medium:
          type: integer
        high:
          type: integer
        veryHigh:
          type: integer
        critical:
          type: integer
      required:
        - low
        - medium
        - high
        - veryHigh
        - critical
    TrendPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
        label:
          type: string
      required:
        - date
        - value
...
    RecentQuery:
      type: object
      properties:
        id:
          type: string
        nrc:
          type: string
        customerName:
          type: string
        score:
          type: number
        rating:
          type: string
        queriedAt:
          type: string
          format: date-time
        decision:
          type: string
          enum: [approved, declined, referred]
      required:
        - id
        - nrc
        - customerName
        - score
        - rating
        - queriedAt
        - decision
    FetchDataRequest:
      type: object
      properties:
        nrc:
          type: string
      required:
        - nrc
    BankDataResponse:
      type: object
      properties:
        nrc:
          type: string
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/BankAccount"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        source:
          type: string
      required:
        - nrc
        - accounts
        - transactions
        - source
    MnoDataResponse:
      type: object
      properties:
        nrc:
          type: string
        mobileMoneyBalance:
          type: number
        averageMonthlyTransactions:
          type: number
        totalTransactionVolume:
          type: number
        accountAge:
          type: integer
          description: Account age in months
        airtime:
          type: number
        source:
          type: string
      required:
        - nrc
        - mobileMoneyBalance
        - averageMonthlyTransactions
        - totalTransactionVolume
        - accountAge
        - airtime
        - source
    MfiDataResponse:
      type: object
      properties:
        nrc:
          type: string
        loans:
          type: array
          items:
            $ref: "#/components/schemas/LoanRecord"
        repaymentRate:
          type: number
        source:
          type: string
      required:
        - nrc
        - loans
        - repaymentRate
        - source
    BankAccount:
      type: object
      properties:
        id:
          type: string
        accountNumber:
          type: string
        type:
          type: string
          enum: [savings, current, fixed_deposit]
        balance:
          type: number
        currency:
          type: string
        bank:
          type: string
        openedAt:
          type: string
          format: date-time
      required:
        - id
        - accountNumber
        - type
        - balance
        - currency
        - bank
        - openedAt
    Transaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        category:
          type: string
      required:
        - id
        - type
        - amount
        - currency
        - description
        - date
        - category
...