# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zambia Multi-Tenant Enterprise Credit Scoring Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface LoginRequest {
  email: string;
  password: string;
  /** Optional tenant code for tenant users */
  tenantCode?: string;
}