# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zambia Multi-Tenant Enterprise Credit Scoring Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface SuccessResponse {
  success: boolean;
  message: string;
}