# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zambia Multi-Tenant Enterprise Credit Scoring Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface TrendPoint {
  date: string;
  value: number;
  label?: string;
}